+++ docs-site/.storybook/main.js
@@ -1,0 +1,11 @@
+module.exports = {
+  stories: ['../stories/**/*.mdx'],
+  addons: ['@storybook/addon-docs'],
+  framework: {
+    name: '@storybook/html',
+    options: {}
+  },
+  docs: {
+    autodocs: false
+  }
+};